APIと処理の分解
画像保存などの処理を別ファイルに移動。
IO待ちでAPIが止まるのを回避するためにバックグラウンド実行に変更。
