+++ pylintrc
@@ -1,0 +1,1 @@
+init-hook="./app"
